After pulling an all-nighter, I made it through the day and even had a successful meeting with someone I'll call my coach: SS. Not THAT SS... I've spent a little too much time dillying and dallying. I have definitely obsessed too much over the finest details in this project. I just so badly want to reealllly understand what I am implementing and WHY.
I tried some cool stuff using counties separately, and may
continue along that path. Not sure how to approach it.
I'll head home and go to sleep for a few hours, and then
awaken with a refreshed mind and complete willingness
to accomplish all that I can in this project.

Godspeed you, reader.